+++ packages/terra-button/src/Button.jsx
@@ -118,7 +118,7 @@ const defaultProps = {
 class Button extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { active: false, focused: false, isSelected: false };
+    this.state = { active: false, focused: false };
     this.handleKeyDown = this.handleKeyDown.bind(this);
     this.handleKeyUp = this.handleKeyUp.bind(this);
     this.handleOnBlur = this.handleOnBlur.bind(this);
@@ -144,7 +144,7 @@ class Button extends React.Component {
     }
   }
 
-  handleClick(event, isSelectable) {
+  handleClick(event) {
     // See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Button#Clicking_and_focus
     // Button on Firefox, Safari and IE running on OS X does not receive focus when clicked.
     // This will put focus on the button when clicked if it is not currently the active element.
@@ -155,15 +155,11 @@ class Button extends React.Component {
     }
 
     if (this.props.onClick) {
-      if (isSelectable) {
-        this.props.onClick(event, this.state.isSelected);
-      } else {
-        this.props.onClick(event);
-      }
+      this.props.onClick(event);
     }
   }
 
-  handleKeyDown(event, isSelectable) {
+  handleKeyDown(event) {
     // Add active state to FF browsers
     if (event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
       this.setState({ active: true });
@@ -179,9 +175,6 @@ class Button extends React.Component {
     // Add focus styles for keyboard navigation
     if (event.nativeEvent.keyCode === KeyCode.KEY_SPACE || event.nativeEvent.keyCode === KeyCode.KEY_RETURN) {
       this.setState({ focused: true });
-      if (isSelectable) {
-        this.setState(prevState => ({ isSelected: !prevState.isSelected }));
-      }
     }
 
     if (this.props.onKeyDown) {
@@ -215,12 +208,9 @@ class Button extends React.Component {
     }
   }
 
-  handleMouseDown(event, isSelectable) {
+  handleMouseDown(event) {
     if (this.props.onMouseDown) {
       this.props.onMouseDown(event);
-    }
-    if (isSelectable) {
-      this.setState(prevState => ({ isSelected: !prevState.isSelected }));
     }
 
     // See https://developer.mozilla.org/en-US/docs/Web/API/HTMLOrForeignElement/focus#Notes
@@ -256,8 +246,8 @@ class Button extends React.Component {
     const isMac = () => navigator.userAgent.indexOf('Mac') !== -1;
     const buttonLabelCx = isMac() ? 'button-label-mac' : 'button-label-win';
 
-    // TODO: `isSelectable` prop is used for fusion pass through passivity and should be removed after Fusion Phase2 release.
-    const { isSelectable } = customProps;
+    // TODO: `isSelectable` and `isSelected` prop is used for fusion pass through passivity and should be removed after Fusion Phase2 release.
+    const { isSelectable, isSelected } = customProps;
 
     const buttonClasses = classNames(
       cx([
@@ -268,7 +258,7 @@ class Button extends React.Component {
         { compact: isCompact },
         { 'is-active': this.state.active && !isDisabled },
         { 'is-focused': this.state.focused && !isDisabled },
-        { 'is-selected': isSelectable && this.state.isSelected && !isDisabled },
+        { 'is-selected': isSelectable && isSelected && !isDisabled },
         theme.className,
       ]),
       customProps.className,
@@ -317,7 +307,7 @@ class Button extends React.Component {
     }
 
     if (isSelectable) {
-      customProps['aria-pressed'] = this.state.isSelected;
+      customProps['aria-pressed'] = isSelected;
     }
 
     let ComponentType = 'button';
@@ -335,12 +325,12 @@ class Button extends React.Component {
         tabIndex={isDisabled ? '-1' : customProps.tabIndex}
         aria-disabled={isDisabled}
         aria-label={ariaLabel}
-        onKeyDown={(event) => { this.handleKeyDown(event, isSelectable); }}
+        onKeyDown={(event) => { this.handleKeyDown(event); }}
         onKeyUp={this.handleKeyUp}
         onBlur={this.handleOnBlur}
         title={buttonTitle}
-        onClick={(event) => { this.handleClick(event, isSelectable); }}
-        onMouseDown={(event) => { this.handleMouseDown(event, isSelectable); }}
+        onClick={(event) => { this.handleClick(event); }}
+        onMouseDown={(event) => { this.handleMouseDown(event); }}
         onFocus={this.handleFocus}
         href={href}
         ref={refCallback}