+++ packages/terra-dropdown-button/tests/jest/DropdownButton.test.jsx
@@ -3,9 +3,20 @@ import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 import { IntlProvider } from 'react-intl';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
 import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { v4 as uuidv4 } from 'uuid';
 import translationsFile from '../../translations/en.json';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
 import DropdownButton, { Item, Variants } from '../../src/DropdownButton';
+
+let mockSpyUuid;
+beforeAll(() => {
+  mockSpyUuid = jest.spyOn(uuidv4, 'v4').mockReturnValue('00000000-0000-0000-0000-000000000000');
+});
+
+afterAll(() => {
+  mockSpyUuid.mockRestore();
+});
 
 describe('Dropdown Button', () => {
   it('should render a default dropdown type', () => {
@@ -87,7 +98,7 @@ describe('Dropdown Button', () => {
   it('correctly applies the theme context className', () => {
     const wrapper = mountWithIntl(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
-        <DropdownButton label="Primary Option">
+        <DropdownButton label="Primary Option" id="dropDown">
           <Item label="1st Option" onSelect={() => {}} />
         </DropdownButton>
       </ThemeContextProvider>,