+++ packages/terra-button/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* Fixed
-  * Fixed lint errors
+* Changed
+  * Added support to programmatically set selection state for Fusion-Button
 
 ## 3.69.0 - (October 23, 2023)