+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,7 +4,10 @@
 
 * Added
   * Added about page for `TagList`.
-  
+
+* Updated
+  *  Updated example for selection state of Fusion-Button in `terra-button`.
+
 ## 1.47.0 - (October 26, 2023)
 
 * Changed
@@ -32,6 +35,7 @@
 
 * Updated
   * Updated `terra-alert` documentation for custom titles.
+  * Updated `terra-dropdown-button` documentation for testing uuid details.
   * Updated exisiting examples upto a11y standards
 
 * Fixed